# docs/swagger.yaml
definitions:
...
  controllers.AllPhoneResponse:
    properties:
      brand:
        example: Samsung
        type: string
        x-order: "1"
      image_url:
        example: ""
        type: string
        x-order: "5"
      model:
        example: Galaxy
        type: string
        x-order: "2"
      phone_id:
        example: 1
        type: integer
        x-order: "0"
      price:
        example: 10000
        type: integer
        x-order: "4"
      release_date:
        example: "2023-11-11T00:00:00+07:00"
        type: string
        x-order: "3"
    type: object
...
  controllers.CreatePhoneInput:
    properties:
      brand:
        example: Samsung
        type: string
        x-order: "1"
      image_url:
        example: ""
        type: string
        x-order: "5"
      model:
        example: Galaxy
        type: string
        x-order: "2"
      price:
        example: 10000
        type: integer
        x-order: "4"
      release_date:
        example: "2023-01-01T00:00:00Z"
        type: string
        x-order: "3"
    type: object
  controllers.CreateReviewInput:
    properties:
      rating:
        example: 1
        type: integer
      review_text:
        example: sample review text
        type: string
    required:
    - rating
    - review_text
    type: object
  controllers.CreatedPhoneResponse:
    properties:
      message:
        example: phone created successfully
        type: string
        x-order: "0"
      phone:
        properties:
          brand:
            example: Samsung
            type: string
            x-order: "1"
          image_url:
            example: ""
            type: string
            x-order: "5"
          model:
            example: Galaxy
            type: string
            x-order: "2"
          phone_id:
            example: 1
            type: integer
            x-order: "0"
          price:
            example: 10000
            type: integer
            x-order: "4"
          release_date:
            example: "2023-01-01T00:00:00Z"
            type: string
            x-order: "3"
        type: object
        x-order: "1"
    type: object
...
  controllers.PhoneByIDResponse:
    properties:
      brand:
        example: Samsung
        type: string
        x-order: "1"
      image_url:
        example: ""
        type: string
        x-order: "5"
      model:
        example: Galaxy
        type: string
        x-order: "2"
      phone_id:
        example: 1
        type: integer
        x-order: "0"
      price:
        example: 10000
        type: integer
        x-order: "4"
      release_date:
        example: "2023-11-11"
        type: string
        x-order: "3"
      reviews:
        items:
          $ref: '#/definitions/controllers.ReviewResponse'
        type: array
        x-order: "6"
    type: object
...
  controllers.ReviewResponse:
    properties:
      phone_id:
        example: 1
        type: integer
      rating:
        example: 9
        type: integer
      review_id:
        example: 1
        type: integer
      review_text:
        example: product is nice
        type: string
      total_likes:
        example: 10
        type: integer
      user_id:
        example: 1
        type: integer
    type: object
  models.Comment:
    properties:
      comment_id:
        example: 1
        type: integer
      comment_text:
        example: sample comment text
        type: string
      created_at:
        example: "2030-01-01 00:00:00"
        type: string
      likes:
        items:
          $ref: '#/definitions/models.Like'
        type: array
      review_id:
        example: 1
        type: integer
      user_id:
        example: 1
        type: integer
    type: object
  models.Like:
    properties:
      comment_id:
        example: 1
        type: integer
      created_at:
        example: "2030-01-01 00:00:00"
        type: string
      like_id:
        example: 1
        type: integer
      review_id:
        example: 1
        type: integer
      user_id:
        example: 1
        type: integer
    type: object
  models.Review:
    properties:
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
        x-order: "5"
      created_at:
        example: "2030-01-01 00:00:00"
        type: string
        x-order: "5"
      likes:
        items:
          $ref: '#/definitions/models.Like'
        type: array
        x-order: "6"
      phone_id:
        example: 1
        type: integer
        x-order: "1"
      rating:
        example: 8
        type: integer
        x-order: "3"
      review_id:
        example: 1
        type: integer
        x-order: "0"
      review_text:
        example: sample review text
        type: string
        x-order: "4"
      user_id:
        example: 1
        type: integer
        x-order: "2"
    type: object
info:
  contact: {}
paths:
...
  /phones:
    get:
      description: Get a list of Phones
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controllers.AllPhoneResponse'
            type: array
      summary: List all phones
      tags:
      - Phones
    post:
      description: Create a phone
      parameters:
      - description: the body to create a phone
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.CreatePhoneInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controllers.CreatedPhoneResponse'
      security:
      - Bearer: []
      summary: Create a phone
      tags:
      - Phones
  /phones/{phone_id}:
    get:
      description: Get phone details by ID
      parameters:
      - description: PhoneID
        in: path
        name: phone_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.PhoneByIDResponse'
      summary: Get phone details by ID
      tags:
      - Phones
  /phones/{phone_id}/reviews:
    get:
      description: Get all phone reviews by ID
      parameters:
      - description: PhoneID
        in: path
        name: phone_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Review'
            type: array
      summary: Get all phone reviews by its ID
      tags:
      - Phones
    post:
      description: Create phone review by phone ID
      parameters:
      - description: PhoneID
        in: path
        name: phone_id
        required: true
        type: string
      - description: the body to create phone review
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateReviewInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Review'
      security:
      - Bearer: []
      summary: Create a phone review by phone ID
      tags:
      - Phones
...
securityDefinitions:
  Bearer:
    description: 'Type "Bearer" followed by a space and JWT token (for example: "Bearer
      eyxyz").'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"